Teacher-Backend/pyproject.toml
[project]
name = "teacher-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "faker>=37.1.0",
    "firebase-admin>=6.8.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.168.0",
    "google-cloud-storage>=2.19.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-google-vertexai>=2.0.21",
    "langsmith>=0.3.35",
    "mysql-connector-python>=9.3.0",
    "pdfminer-six>=20250416",
    "pybase64>=1.4.1",
    "pyjwt>=2.10.1",
    "pymssql>=2.3.4",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
